requirements.txt
openai==1.99.1
openpyxl==3.1.5
python-dotenv==1.1.1
streamlit==1.48.0
mistralai==1.9.3
google-genai==1.29.0